Extract KeyLog into component file
